backend/requirement.txt
Django>=4.2,<5.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
openpyxl>=3.1.0